Fix signed comparisons in utils/common.

+++ mp/src/utils/common/cmdlib.cpp
@@ -808,14 +808,14 @@ FileHandle_t SafeOpenRead( const char *filename )
 
 void SafeRead( FileHandle_t f, void *buffer, int count)
 {
-	if ( g_pFileSystem->Read (buffer, count, f) != (size_t)count)
+	if ( (size_t)g_pFileSystem->Read (buffer, count, f) != (size_t)count)
 		Error ("File read failure");
 }
 
 
 void SafeWrite ( FileHandle_t f, void *buffer, int count)
 {
-	if (g_pFileSystem->Write (buffer, count, f) != (size_t)count)
+	if ((size_t)g_pFileSystem->Write (buffer, count, f) != (size_t)count)
 		Error ("File write failure");
 }
